Add files via upload
The setwd('C:/Users/[your username]/OneDrive/Documents') can be changed to the place where the CSV file is kept in your computer or put your username minus the square brackets. The CSV file is uploaded along with the code and the presentation. There two libraries used ggplot2 and moments which can downloaded from Tools -> Install Packages if you are using R studio.